apps/vqa-service/app.py
from flask import Flask, request
from flask_restx import Api, Resource, fields
from PIL import Image
import io
import werkzeug

app = Flask(__name__)
api = Api(app, version='1.0', title='VQA Service API',
          description='A simple API for Visual Question Answering')

# Define namespace
ns = api.namespace('api', description='VQA operations')

# Define models for Swagger documentation
element_model = api.model('Element', {
    'type': fields.String(required=True, description='Type of UI element'),
    'position': fields.List(fields.Integer, required=True, description='Position coordinates [x, y]')
})

response_model = api.model('Response', {
    'page_type': fields.String(required=True, description='Type of the page'),
    'confidence': fields.Float(required=True, description='Confidence score'),
    'elements': fields.List(fields.Nested(element_model), required=True, description='List of UI elements')
})

# File upload parser
upload_parser = api.parser()
upload_parser.add_argument('image', 
                         type=werkzeug.datastructures.FileStorage, 
                         location='files',
                         required=True, 
                         help='Image file to analyze')

@ns.route('/query-page-intent')
class QueryPageIntent(Resource):
    @ns.expect(upload_parser)
    @ns.response(200, 'Success', response_model)
    @ns.response(400, 'Validation Error')
    def post(self):
        """Submit an image for page intent analysis"""
        args = upload_parser.parse_args()
        image_file = args['image']
        
        # Validate file type
        if not image_file.filename or '.' not in image_file.filename:
            api.abort(400, 'Invalid file format')
        
        try:
            # Read and validate image
            image_bytes = image_file.read()
            image = Image.open(io.BytesIO(image_bytes))
            
            # TODO: Add your image processing logic here
            # For now, return a dummy response
            response = {
                'page_type': 'form',
                'confidence': 0.95,
                'elements': [
                    {'type': 'input_field', 'position': [100, 200]},
                    {'type': 'button', 'position': [300, 400]}
                ]
            }
            
            return response
        
        except Exception as e:
            api.abort(400, str(e))

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000, debug=True)
