apps/vqa-service/requirements.txt
flask==3.0.0
Pillow==10.1.0
python-dotenv==1.0.0
werkzeug==3.0.1
flask-restx==1.3.0
